Add files via upload
The objective of the game is when
the snake see a fruit it will eat it and in return the size of the snake grows. If the snake’s
body touches the border of the game window then the player will lose and game is over. The
player also has the ability to either lose the game or simply quit it. There are also some
chances of innovations as it can be turned into a multiplayer game which is still in progress.